Hit 30 fps at 6x360: int32 colour pipeline, accurate host pacing
30 fps was the hard target and 6x360 all-ACTIVE managed 27.8 on the board
and ~24 through the driver. Four fixes, measured individually.

The dominant one is arithmetic width. oklchQ14ToRgb8 did ~20 int64
multiplies per pixel, and the RP2040 is a Cortex-M0+ with no 64-bit
multiply at all -- every one became an __aeabi_lmul call, which measured
~12 us/pixel (~1500 cycles) for what should be a handful of MULS. The
OKLab->LMS matrix, the a/b projection and cube_q14 narrow to int32; the
LMS->RGB accumulator deliberately does not, since its worst case reaches
~2.04e9 against a 2.15e9 limit. 6x360 went 27.8 -> 38.7 fps, 8x360
19.6 -> 29.3. Bit-identical, which the golden vectors confirm; a new host
check derives the overflow bounds from the actual matrices, so retuning a
coefficient or widening a quantization level fails the test rather than
silently truncating colour.

SerialDriver.run paced with time.sleep(), whose Windows granularity is
~15.6ms -- it capped the host near 24 fps no matter how fast the board ran.
It now requests a 1ms timer period and spins the final 2ms.

_route COBS-decoded entire frames to read a 13-byte header; it now decodes
the header only, via cobs_decode_header.

HELLO was never receivable: setup() writes it before USB-CDC enumerates, so
every session burned the full hello_timeout for nothing. The board repeats
HELLO until its first frame arrives. open() went 2.07s -> 0.10s, and the
handshake does something for the first time.

End to end through SerialDriver, zero RESYNC and zero window stalls:
6x360 all-ACTIVE 29.91 fps, 6x360 every-other-INTERPOLATED 30.00,
8x360 all-ACTIVE 29.38.

Also resolves the flow-control question left open last session. The single
overdrive failure seen with the window in place was the double-buffering
experiment, since reverted; on the shipped firmware that test has now run
clean five consecutive times (~105s, ~150k frames) with the board alive
after each. Both attempts at overlapping DMA with CPU work measured worse
and are recorded in the README so they are not retried blindly.

firmware/tools/bench.py replaces the throwaway harness that lived in a temp
directory and was lost. It keeps the two measurement traps that produced
wrong numbers before -- sleep-based pacing, and buffer wraps that truncate
a frame mid-flight -- out of reach by construction.

pytest 199 passed / 0 skipped, C++ golden conformance bit-exact with 4
robustness checks, mypy and black clean.

Known open: budget_for_baud still sizes the per-frame budget from link rate
rather than device throughput, so the default 5333 B/frame yields 12.4 fps;
callers need an explicit budget_bytes of ~800 until that is fixed.

+++ firmware/scorpio/lib/lumicodec/lumicodec.cpp
@@ -343,12 +343,30 @@ inline int32_t cosInterp_q14(int32_t h_88) {
   return a + (((b - a) * frac) >> 8);
 }
 
+// 32-bit where the operand range provably allows it. The RP2040 is a
+// Cortex-M0+: 32x32->32 is a single-cycle MULS, but it has no 64-bit
+// multiply at all, so every int64 product becomes an __aeabi_lmul call.
+// That dominated the per-pixel cost (measured ~12 us/pixel, ~1500 cycles).
+//
+// Bounds over the whole quantized input space -- qL in 0..63 so l_q14 <=
+// 16384, qC in 0..31 so c_q14 <= 6554, cos/sin in +-16384, and INTERPOLATED
+// values are blends of two such and so no larger:
+//
+//   x*x in cube_q14   <= 25400^2       = 6.5e8   (int32 max 2.15e9)
+//   xx*x in cube_q14  <= 39400*25400   = 1.0e9
+//   gammaEncode       <= 124000*255    = 3.2e7
+//
+// None can overflow, so results are bit-identical to the int64 form -- which
+// the golden-vector conformance test (spec 11.9) verifies directly. The
+// LMS->RGB accumulator in oklchQ14ToRgb8 is deliberately NOT converted: its
+// worst case reaches ~2.04e9 against a 2.15e9 limit, too little margin to
+// call safe.
 inline int32_t cube_q14(int32_t x) {
-  int64_t xx = (static_cast<int64_t>(x) * x) >> 14;
+  int32_t xx = static_cast<int32_t>((x * x) >> 14);
   return static_cast<int32_t>((xx * x) >> 14);
 }
 
-inline uint8_t gammaEncode(int64_t linear_q14, uint8_t scale1, uint8_t scale2) {
+inline uint8_t gammaEncode(int32_t linear_q14, uint8_t scale1, uint8_t scale2) {
   if (linear_q14 < 0) linear_q14 = 0;
   // brightness and per-channel correction in linear space (spec §8.4.3)
   linear_q14 = (linear_q14 * scale1) >> 8;
@@ -364,21 +382,27 @@ void oklchQ14ToRgb8(int32_t l_q14, int32_t c_q14, int32_t h_88,
   buildTables();
   int32_t cosH = cosInterp_q14(h_88);
   int32_t sinH = cosInterp_q14(((64 << 8) - h_88) & 0xFFFF);  // sin x = cos(64-x)
-  int32_t a = static_cast<int32_t>((static_cast<int64_t>(c_q14) * cosH) >> 14);
-  int32_t b = static_cast<int32_t>((static_cast<int64_t>(c_q14) * sinH) >> 14);
+  // c_q14 <= 6554 and |cos| <= 16384, so the product peaks near 2^26.7 --
+  // int32 with room to spare. See the bounds note above cube_q14().
+  int32_t a = (c_q14 * cosH) >> 14;
+  int32_t b = (c_q14 * sinH) >> 14;
 
   int32_t lms[3];
   for (int i = 0; i < 3; i++) {
-    int64_t acc = static_cast<int64_t>(M_L2LMS[i * 3]) * l_q14 +
-                  static_cast<int64_t>(M_L2LMS[i * 3 + 1]) * a +
-                  static_cast<int64_t>(M_L2LMS[i * 3 + 2]) * b;
-    lms[i] = cube_q14(static_cast<int32_t>(acc >> 14));
+    // Worst row peaks near 4.2e8 against an int32 limit of 2.15e9.
+    int32_t acc = M_L2LMS[i * 3] * l_q14 +
+                  M_L2LMS[i * 3 + 1] * a +
+                  M_L2LMS[i * 3 + 2] * b;
+    lms[i] = cube_q14(acc >> 14);
   }
   for (int i = 0; i < 3; i++) {
+    // Stays 64-bit: this one peaks near 2.04e9, too close to the int32
+    // limit to be provably safe.
     int64_t acc = static_cast<int64_t>(M_LMS2RGB[i * 3]) * lms[0] +
                   static_cast<int64_t>(M_LMS2RGB[i * 3 + 1]) * lms[1] +
                   static_cast<int64_t>(M_LMS2RGB[i * 3 + 2]) * lms[2];
-    out[i] = gammaEncode(acc >> 14, brightness, correction[i]);
+    out[i] = gammaEncode(static_cast<int32_t>(acc >> 14), brightness,
+                         correction[i]);
   }
 }
